+++ src/main/java/domain/piece/category/Cannon.java
@@ -2,7 +2,7 @@ package domain.piece.category;
 
 import domain.direction.Directions;
 import domain.piece.Piece;
-import domain.piece.Position;
+import domain.spatial.Position;
 
 public class Cannon extends Piece {
