+++ src/main/java/domain/spatial/Position.java
@@ -1,4 +1,4 @@
-package domain.piece;
+package domain.spatial;
 
 import java.util.Objects;
