+++ src/test/java/domain/BoardTest.java
@@ -9,8 +9,8 @@ import domain.piece.Piece;
 import domain.piece.PieceColor;
 import domain.piece.PieceInit;
 import domain.piece.Pieces;
-import domain.piece.Position;
 import domain.piece.category.Soldier;
+import domain.spatial.Position;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
