+++ src/test/java/domain/direction/DirectionTest.java
@@ -2,7 +2,7 @@ package domain.direction;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.piece.Position;
+import domain.spatial.Position;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
