+++ src/test/java/domain/direction/DirectionsTest.java
@@ -1,11 +1,10 @@
 package domain.direction;
 
-import domain.piece.Position;
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
-
 import static org.assertj.core.api.Assertions.assertThat;
+
+import domain.spatial.Position;
+import java.util.List;
+import org.junit.jupiter.api.Test;
 
 class DirectionsTest {
