+++ src/test/java/domain/piece/PieceTest.java
@@ -12,6 +12,7 @@ import domain.piece.category.Guard;
 import domain.piece.category.Horse;
 import domain.piece.category.King;
 import domain.piece.category.Soldier;
+import domain.spatial.Position;
 import java.util.List;
 import org.junit.jupiter.api.Test;
