+++ src/test/java/domain/piece/PiecesTest.java
@@ -9,6 +9,7 @@ import domain.piece.category.Cannon;
 import domain.piece.category.Horse;
 import domain.piece.category.King;
 import domain.piece.category.Soldier;
+import domain.spatial.Position;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
@@ -135,7 +136,7 @@ class PiecesTest {
 
         // when & then
         pieces.isCannonByPosition(position1);
-        
+
         // then
         assertAll(
                 () -> assertThat(pieces.isCannonByPosition(position1)).isTrue(),
